+++ custom_components/wnhf/executions/generic_real.py
@@ -92,7 +92,7 @@ class GenericExecutionEngine:
     """
 
     API_VERSION = "1.0"
-    VERSION = "1.7-rc5"
+    VERSION = "1.8-rc6"
     ENABLED_ACTIONS = REAL_EXECUTION_ENABLED_ACTIONS
 
     def __init__(